# infra/terraform/modules/apim/openapi/gateway.yaml
openapi: 3.0.0
info:
  title: Carpeta Ciudadana Gateway API
  description: API Gateway for Carpeta Ciudadana operator
  version: 1.0.0
  contact:
    name: Carpeta Ciudadana Team
    email: support@carpeta-ciudadana.com
servers:
  - url: https://{apim-name}.azure-api.net/api
    description: API Management Gateway
paths:
  /citizens:
    post:
      summary: Register a new citizen
      description: Register a new citizen in the system
      operationId: register-citizen
      tags:
        - Citizens
      x-azure-backend:
        serviceUrl: "{{citizen-backend}}"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CitizenCreate'
      responses:
        '201':
          description: Citizen registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Citizen'
        '400':
          description: Bad request
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error

    get:
      summary: List citizens
      description: Get list of citizens
      operationId: list-citizens
      tags:
        - Citizens
      x-azure-backend:
        serviceUrl: "{{citizen-backend}}"
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of citizens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Citizen'
  /citizens/{citizen_id}:
    get:
      summary: Get citizen by ID
      description: Get citizen information by ID
      operationId: get-citizen
      tags:
        - Citizens
      x-azure-backend:
        serviceUrl: "{{citizen-backend}}"
      parameters:
        - name: citizen_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Citizen information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Citizen'
        '404':
          description: Citizen not found
...
components:
  schemas:
    CitizenCreate:
      type: object
      required:
        - email
        - full_name
        - document_number
        - document_type
      properties:
        email:
          type: string
          format: email
        full_name:
          type: string
        document_number:
          type: string
        document_type:
          type: string
          enum: [CC, CE, PA, TI]
    Citizen:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        full_name:
          type: string
        document_number:
          type: string
        document_type:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...